// src/app/weather/weather-data.interface.ts
export interface WeatherData {
  city: string;
  temperature: number;
  tempMin:number;
  tempMax:number;
  feelsLikeTemp: number;
  humidity: number;
  pressure: number;
  summary: string;
  iconURL: string;
  wind:number ;
  description:string;
  mainDes:string;
  sunRise:string;
  sunSet:string;
}